+++ lib/operations/models.ex
@@ -8,7 +8,8 @@ defmodule OpenAi.Models do
   @doc """
   Delete a fine-tuned model. You must have the Owner role in your organization to delete a model.
   """
-  @spec delete_model(String.t(), keyword) :: {:ok, OpenAi.Model.DeleteResponse.t()} | :error
+  @spec delete_model(String.t(), keyword) ::
+          {:ok, OpenAi.Model.DeleteResponse.t()} | {:error, OpenAi.Error.error()}
   def delete_model(model, opts \\ []) do
     client = opts[:client] || @default_client
 
@@ -25,7 +26,8 @@ defmodule OpenAi.Models do
   @doc """
   Lists the currently available models, and provides basic information about each one such as the owner and availability.
   """
-  @spec list_models(keyword) :: {:ok, OpenAi.Model.ListResponse.t()} | :error
+  @spec list_models(keyword) ::
+          {:ok, OpenAi.Model.ListResponse.t()} | {:error, OpenAi.Error.error()}
   def list_models(opts \\ []) do
     client = opts[:client] || @default_client
 
@@ -42,7 +44,8 @@ defmodule OpenAi.Models do
   @doc """
   Retrieves a model instance, providing basic information about the model such as the owner and permissioning.
   """
-  @spec retrieve_model(String.t(), keyword) :: {:ok, OpenAi.Model.t()} | :error
+  @spec retrieve_model(String.t(), keyword) ::
+          {:ok, OpenAi.Model.t()} | {:error, OpenAi.Error.error()}
   def retrieve_model(model, opts \\ []) do
     client = opts[:client] || @default_client